Completely revamped the UI to more closely match the CF6.1 design
This CR6Community Edition of DGUS Reloaded is intended to be "easy" for current users of the CF6.1 firmware to navigate and learn.  It also takes advantage of the aesthetics and use-case driven UI design work that went into the CF6.1 project deliverables.

Some of the Klipper back-end functionality is still not properly understood and working, so this version of the front end does not yet operate as intended.  It is being made available in this condition as a marked improvement over v0.2-beta.  User review feedback and design input is welcome, in the Discussions section of either repo.