Animate print time icon | Change Filament while Paused | Update Release Notes
This completes the UI update for 0.3.5.

The Printing Status/Paused/FInished screens have been modified to add animation to the timer icon.  Hoping to remove the progress bar if new features need the space.

Added the ability to change filament while a print is paused. (NOTE: May still not work if the printer decides it is "Busy", but tested ok in my trials.)