Nudged Gcode_Z_Offset field slightly to right. Removed old HZK file
Took this opportunity to nudge the GCode__Offset field slightly to the right & off the frame border on all screens where it appears.
Also removed the original HZK file, since the one that is new at v3.3 is now the required font.  Updated the readme in the one-time DWIN files folder, to remove reference to renaming HZK before flashing (no longer required)