Shift Gcode Z Offset display to right
Display was center-justified in DGUSTool, but showed as left-justified on the screen and did not move right to reveal the minus sign when negative values were entered..
Moved the Up Arrow icon to the right, in case it was masking the data value + changed the justification to right-justified, then shifted the display field left a few points.
Data display is now visible whether +ve, -ve or 0.