Add Enable/Disable RunoutSensor control to Setup page
Creates a second control which may become out of synch with Mainsail and the actual printer state. (Other is LED On/Off).

Mixed feelings about releasing in this configuration, but not yet able to reprogram in a way that senses and corrects this conflict.  Is on to-do list, to fix both, when able.