Minor bug fixes and enhancements
Raised the version to 0.4.4.
Moved the version number on the Home page and added an information icon, to make it more obvious that there is an Information page and to move the Information button away from the Emergency Shutdown button.
Removed GUI Version and DGUS2 Version from the Information page, since neither value makes sense to the user.
Added t5uid1 version number, which is now branded into vars_out.cfg in Klipper Component version 1.3.5 and subs
Moved the Klipper version number in memory and extended it to 21 characters long, from 16, to display the full value.