Overhauled the Automatic Bed Leveling page
Added colour-coding to the  bed mesh, with a user-defined threshold for what is deemed "green"

Reworked the ABL workflow, to allow repeatedly pressing the LOAD Profile button to pagie through the existing bed mesh profiles and selecting Run ABL to generate a new bed_mesh to replace the one currently displayed.  (Requires presets.cfg from the Klipper component, which includes a user-modifiable list of the bed_mesh profile names and the macros that generate them.)

Added a RESTART KLIPPER option, if the user needs to scrap the new profile and keep the prior one. (Otherwise, Klipper will "lie in wait" for the next SAVE_CONFIG command, to save the new profile into printer.cfg.)

Added feedback to the Probing screen, to give the user some confidence that they know what the printer is doing, when it runs the ABL macros.

Reworked the SETUP screen, to create some space for future functionality.  Removed the Screen Brightness Full/Min buttons to create that space.

Added new screen bitmaps for a future Extruder Calibration function (not implemented at this release)
Removed the old Heat Presets popup menus, made redundant at v0.4.5 (Moved to the Other Graphics folder, just in case...)