Repaired files broken at v0.4.7
Again, GitHub Desktop seems to have betrayed my trust.  Had to recreate all of the screens  numbered higher than 020.
Not sure whether it is only my system that is screwed-up or whether the online release 0.4.7 is also busted.

If no one else complains, I will hope for the best, but to be able to do more development, I need to recreate my baseline...