Modified T5LCFG file to silence the touch beeper.  Programmed the buttons to beep when operated.
I am trying to improve the UI by debouncing some of the keys.   I find it very confusing when the touchscreen beeps because I touched it, but the button does not operate until after the programmed delay.  It was less confusing to silence the hardware and just to beep when the software says the button action has begun.
On the other hand, now the UI does not beep when a button causes a popup screen, nor do keypad keystrokes beep.  Those controls only beep when the final action is taken on the popup screen.  Some may find that behaviour change to be disturbing, instead.
I will release v.0.4.8 this way, and will consider whether to revert or modify the UI based on regular use of the printer.